requirements.txt
opencv_contrib_python==4.5.5.64
Pillow==9.3.0
scikit_image==0.19.2
skimage==0.0
streamlit==1.5.1
